Merged PR 28034: Azure resources for Defra WebOps to target
Azure resources for Defra WebOps to target

Added to the environment configuration:

 - A DNS Zone for each environment that DDTS can delegate to for the corresponding subdomain.
 - Added a KeyVault for SSL certificate storage.

Added `subscriptionId` property to inputs and outputs so that _environment_ and _hub_ can be in different subscriptions.

Also added Private DNS Zones for Cosmos DB, Key Vault and Service Bus endpoints and configured these as Private DNS Zone groups for the associated Private Endpoints.