Merged PR 28189: Added new pipeline 'build-release-apps'
the pipeline uses 2 parameters - for environment and for app name (ex. "dev" and "waste-tracking-service") that needs to be selected before pipeline runs.
the previous variables have been reworked to use parameter selection where necessary.
pipeline tested in build runs
https://dev.azure.com/defragovuk/WTS%20-%20Waste%20Tracking%20Service/_build/results?buildId=259344&view=results
https://dev.azure.com/defragovuk/WTS%20-%20Waste%20Tracking%20Service/_build/results?buildId=259357&view=results

old pipelines will be deleted once this is merged to main and approved