Merged PR 30911: Waste Collection Date fixes
Waste Collection Date fixes

Fixed the invalid schema that had broken the _annex-vii_ service. Fixed the failing test and added additional tests to ensure the schema behaves as expected.

Also:

 - Switched from [legacy](https://momentjs.com/docs/) Moment to [date-fns](https://date-fns.org/) and used the provided function for calculating business days.
 - Removed business-rule validation from _waste-tracking-gateway_ route and added this to _annex-vii_ service (and mock backend). Added tests for these.
 - Removed commented-out code, unused code and import.

Waste collection date _Started_ type was incompatible with schema in that _value_ is optional in schema. Discussion with JK noted that _Started_ case is unnecessary, so delete this.

Related work items: #150377