Merged PR 40567: Update repository classes to move methods to sub-classes
BaseRepository methods createTemplateFromSubmission and createSubmissionFromTemplate don't need to be in the BaseRepository.
Which means we don't need the container names to be in the BaseRepository and therefore no longer generic.
Update to move these methods into their appropriate repository classes and rename the attributes to draftContainerName and templateContainerName.
And to keep inline with current repository method names, I'm going to rename the methods createTemplateFromDraft and createDraftFromTemplate.
Also generic method copyCarriers renamed to copyCarriersNoTransport to be more precise.

Related work items: #227777