Merged PR 46145: Verified JWT signature and IDP parameters
Verified JWT signature and IDP parameters

Verified signature of JWT before consuming payload, along with _iss_ and _aud_ claims.

Switched to using [hapi-auth-jwt2](https://github.com/dwyl/hapi-auth-jwt2) which seems to have good support for this; _jwt-js-decode_ doesn't appear to support asymmetric keys.

[jwks-rsa](https://github.com/auth0/node-jwks-rsa) used to obtain public keys from JWKS endpoint.

Needed to make _validateToken_ function compatible and attempted to simplify this in the process. Moved regex and JDT schema compilation outside of function so this only needs to happen once.

Replaced use of _hapi-auth-bearer-token_ in tests with just a mock auth-strategy that just directly sets the credential.

_waste-tracking-gateway_ service now requires `DCID_WELLKNOWN` and `DCID_CLIENT_ID` variables from library.