Merged PR 46153: Makeshift restriction of users
Makeshift restriction of users

Allowed a `ALLOWED_USERS` environment variable to be set with a semicolon-delimited list of DCID _uniqueReference_ values.

If this variable is set and the user isn't in the list, a _403_ is returned which the frontend interprets as the user not being part of the Private Beta.