Merged PR 48489: Client and container build for limited-audience service
Client and container build for limited-audience service

Added a Dapr client for _limited-audience_ service. Also:

 - A tracing module for this service.
 - Another method to the service that the gateway can use to register legacy user-references.
 - An `nx container` target for this service.

With regard to the Dockerfile, there was some difficulty in getting `npm ci` to accept the package-lock file generated by the Webpack build. This appears to be working by switching to the more modern style of webpack.config.json, although I'm unsure why this behaves differently to other services.