Merged PR 49617: Creation of record in submissions container upon draft declaration
Creation of a record in the submissions container once a draft declaration has been signed.
The submission record will have the same details as the draft except the statuses related to each of the sections.
Requires the environment variable COSMOS_SUBMISSIONS_CONTAINER_NAME.
Any updates to the draft after it has attained the COMPLETED status (SubmittedWithEstimates, SubmittedWithActuals, or UpdatedWithActuals) will be propagated to the submission.

Also, updates the Waste Quantity to include the unit, which for bulk will be Tonnes or Cubic Metres, and for small waste - Kilograms and Litres

Related work items: #301538, #302019