Merged PR 51977: Update gitignore file to be compatible with Visual Studio
to update the gitignore file to be more compatible with visual studio full. I remark that the .vs folder is not ignored today and the change in this folder are indicate as changed. Other change that i bring to this file is to ignore the future migrations.json file because this file is not required on the git