Merged PR 55369: 344384 GLW Multiples Submission Error
Adds an error page displayed to the user when the API call to finalise the multiples submission process fails, including an email address the user should contact if they are having errors with their account.

Adds improvement to disable the button which calls the `finalize` endpoint to prevent user from calling the API multiple times in a single journey when their network connection is slow.

Related work items: #344384