DWT-339 Site Waste Authorisation number [Array] (#66)
* Rename and restructure permit fields in Receipt API schema

Revised the `authorisations` field to `authorisationNumbers` and simplified its structure to an array of strings. Added a new `regulatoryPositionStatements` field to handle an array of positive integers for regulatory position statements. Updated descriptions and examples accordingly for clarity and consistency.

* Add validation for 'authorisationNumbers' and remove RPS min value

Enforced a minimum of one item in the 'authorisationNumbers' array to ensure proper validation. Removed the minimum value constraint for RPS numbers to allow greater flexibility.

* Add `authorisationNumbers` as a required field in API spec

Updated the Receipt API specification to include `authorisationNumbers` as a mandatory field. Clarified the description to indicate that at least one authorisation number is required for waste acceptance. This ensures stricter validation and compliance with site authorisation requirements.